requirements.txt
Django==4.0.4
celery==5.3.1
djangorestframework==3.14.0
environs==9.5.0
redis==4.6.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.3
Pillow==7.0.0